requirements.txt
aiogram==3.10.0
APScheduler==3.10.4
feedparser==6.0.11
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
